+++ sdk/resourcemanager/azure-resourcemanager-compute/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Supported `withPlan()` for `VirtualMachineScaleSet` during create.
 - Supported `getNetworkInterfaceByInstanceIdAsync()` in `VirtualMachineScaleSet`.
 - Supported `getNetworkInterfaceAsync()` in `VirtualMachineScaleSetVM`.
+- Supported `orchestrationMode()` in `VirtualMachineScaleSet`.
 
 ## 2.9.0 (2021-10-21)